Revised matmul() to also delegate to matmul_raw() and fixed various issues with gradient calculations, batching, and Dense layer input flattening (added Flatten parameter to Dense layer data - updated .nnn file format)

+++ NNN-Solution/NNN/DataContainers.h
@@ -5,26 +5,26 @@
 // Container struct for convolution geometric data.
 struct ConvGeometry
 {
-	int batches = 0;
-	int spatial_rank = 0;
-	int input_channels = 0;
-	int filter_count = 0;
+	size_t batches = 0;
+	size_t spatial_rank = 0;
+	size_t input_channels = 0;
+	size_t filter_count = 0;
 
 	std::vector<int> input_dims;
 	std::vector<int> kernel_dims;
 	std::vector<int> out_dims;
 
 	std::vector<int> input_strides;
-	std::vector<int> out_spatial_strides;
-	std::vector<int> kernel_spatial_strides;
+	std::vector<size_t> out_spatial_strides;
+	std::vector<size_t> kernel_spatial_strides;
 
-	std::vector<int> input_kernel_offset;
-	std::vector<int> kernel_kernel_offset;
+	std::vector<size_t> input_kernel_offset;
+	std::vector<size_t> kernel_kernel_offset;
 
-	int out_spatial_size = 0;
-	int kernel_spatial_size = 0;
-	int kernel_volume_size = 0;
+	size_t out_spatial_size = 0;
+	size_t kernel_spatial_size = 0;
+	size_t kernel_volume_size = 0;
 
-	int im2col_rows = 0;
-	int im2col_cols = 0;
+	size_t im2col_rows = 0;
+	size_t im2col_cols = 0;
 };
